backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib
PyJWT
openpyxl
msoffcrypto-tool
pytest
httpx
email-validator
